src/test_driver/mbed-functional/requirements.txt
appdirs==1.4.4
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
fasteners==0.15
future==0.18.2
idna==2.9
importlib-metadata==1.6.1
intelhex==2.2.1
junit-xml==1.9
lockfile==0.12.2
lxml==4.5.1
mbed-flasher==0.10.1
mbed-ls==1.7.11
mbed-os-tools==0.0.14
monotonic==1.5
more-itertools==8.3.0
packaging==20.4
pathlib2==2.3.5
pluggy==0.13.1
prettytable==0.7.2
py==1.8.1
pyparsing==2.4.7
pyserial==3.4
pytest==5.4.3
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
wcwidth==0.2.4
zipp==1.2.0